+++ api/controllers/api/tracks.py
@@ -562,12 +562,13 @@ def artwork(username, trackslug):
     if artwork_size > Reel2bitsDefaults.artwork_size_limit:
         return jsonify({"error": "artwork too big, 2MB maximum"}), 413  # Request Entity Too Large
 
-    # Delete old artwork
-    old_artwork = os.path.join(current_app.config["UPLOADED_ARTWORKSOUNDS_DEST"], track.path_artwork())
-    if os.path.isfile(old_artwork):
-        os.unlink(old_artwork)
-    else:
-        print(f"Error: cannot delete old track artwork: {track.id} / {track.artwork_filename}")
+    # Delete old artwork if any
+    if track.artwork_filename:
+        old_artwork = os.path.join(current_app.config["UPLOADED_ARTWORKSOUNDS_DEST"], track.path_artwork())
+        if os.path.isfile(old_artwork):
+            os.unlink(old_artwork)
+        else:
+            print(f"Error: cannot delete old track artwork: {track.id} / {track.artwork_filename}")
 
     # Save new artwork
     artwork_filename = get_hashed_filename(artwork_uploaded.filename)
